add game menu
